+++ src/components/VueTransmit.vue
@@ -16,14 +16,16 @@
 		</div>
 		<slot name="files"
 		      v-bind="fileSlotBindings"></slot>
-		<input type="file"
-		       ref="hiddenFileInput"
-		       :multiple="multiple"
-		       :style="fileInputStyles"
-		       :class="[maxFilesReachedClass]"
-		       :accept="filesToAccept"
-		       :capture="capture"
-		       @change="onFileInputChange">
+		<form :style="formStyles"
+		      ref="uploadForm">
+			<input type="file"
+			       ref="hiddenFileInput"
+			       :multiple="multiple"
+			       :class="[maxFilesReachedClass]"
+			       :accept="filesToAccept"
+			       :capture="capture"
+			       @change="onFileInputChange">
+		</form>
 	</component :is="tag">
 </template>
 
@@ -208,7 +210,7 @@ export default class VueTransmit extends Vue {
 	@Prop({ type: String, default: "You can't upload files of this type." })
 	dictInvalidFileType: string
 	// If the server response was invalid.
-	@Prop({ type: String, default: "Server responded with {{ statusCode }} code." })
+	@Prop({ type: String, default: "Error during upload: {{ statusText }} [{{ statusCode }}]" })
 	dictResponseError: string
 	/**
    * Displayed when the maxFiles have been exceeded
@@ -240,7 +242,7 @@ export default class VueTransmit extends Vue {
 		"Cache-Control": "no-cache",
 		"X-Requested-With": "XMLHttpRequest",
 	}
-	public fileInputStyles: object = {
+	public formStyles: object = {
 		visibility: "hidden !important",
 		position: "absolute !important",
 		top: "0 !important",
@@ -253,6 +255,13 @@ export default class VueTransmit extends Vue {
 		let el = null
 		if (this.$refs.hiddenFileInput instanceof HTMLInputElement) {
 			el = this.$refs.hiddenFileInput
+		}
+		return el
+	}
+	get formEl(): HTMLFormElement {
+		let el = null
+		if (this.$refs.uploadForm instanceof HTMLFormElement) {
+			el = this.$refs.uploadForm
 		}
 		return el
 	}
@@ -327,6 +336,7 @@ export default class VueTransmit extends Vue {
 	}
 	onFileInputChange(): void {
 		this.$emit(Events.AddedFiles, Array.from(this.inputEl.files).map(this.addFile))
+		this.formEl.reset()
 	}
 	addFile(file: File): VTransmitFile {
 		const vtFile = VTransmitFile.fromNativeFile(file)
@@ -566,7 +576,7 @@ export default class VueTransmit extends Vue {
 	handleUploadError(files: VTransmitFile[], xhr: XMLHttpRequest): () => void {
 		const vm = this
 		return function onUploadErrorFn(): void {
-			if (files[0].status !== UploadStatuses.Canceled) {
+			if (files[0].status !== STATUSES.CANCELED) {
 				const message = vm.dictResponseError.replace(hbsRegex, hbsReplacer({ statusCode: xhr.status }))
 				vm.errorProcessing(files, message, xhr)
 			}